app/requirements.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver==2.24.1
chromedriver-binary==112.0.5615.49.0
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.0.0
exceptiongroup==1.1.1
executing==1.2.0
ffmpeg==1.4
ffprobe==0.5
ffprobe-python==1.0.3
filelock==3.11.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.2.0
incremental==22.10.0
ipython==8.12.0
itemadapter==0.8.0
itemloaders==1.0.6
jedi==0.18.2
jmespath==1.0.1
lxml==4.9.2
matplotlib-inline==0.1.6
mutagen==1.46.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==2.0.0
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
Protego==0.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.17
PyDispatcher==2.0.7
pydub==0.25.1
pyee==8.2.2
Pygments==2.14.0
pyOpenSSL==23.1.1
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
redis==4.5.5
requests==2.28.2
requests-file==1.5.1
selenium==4.8.3
service-identity==21.1.0
ShazamAPI==0.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spotipy==2.23.0
sqlparse==0.4.3
stack-data==0.6.2
tldextract==3.4.0
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
wcwidth==0.2.6
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
yt-dlp==2023.3.4
zipp==3.15.0
zope.interface==6.0
